Take tableName pluralisation in to account
When a model is defined without a specified `tableName` sequelize will pluralise model name, unless you have `freezeTablename: true`. This change takes that in to account, and will pluralise the table name in cases where it should.

This was causing issues where `Version` would look for tables that didn't exist, and the codebase was using a mix of models with and without `tablename` and `freezeTablename` properties set consistently.